requirements.txt
flask
flask_cors
seaborn
matplotlib
tqdm
pyyaml
matplotlib
pandas
keras
tensorflow
face_recognition
pillow 
imutils
opencv-python
h5py
easyocr